requirements.txt
beautifulsoup4==4.10.0
cloudscraper==1.2.71
elasticsearch==8.17.0
geopy==2.4.1
numpy==1.21.5
Pillow==11.0.0
playwright==1.43.0
pytesseract==0.3.13
python3_resolver==1.0.3
pytz==2022.1
Requests==2.32.3
requests_tor==1.4
tldextract==5.1.3
